fix: Resolved issue #6 - alterado estrutura do projeto, refactory
