+++ .env-example
@@ -1,9 +1,9 @@
 # LOG CONFIG
-LOG_FILE_ERROR  = log/error.log
-LOG_FILE_ACTIVE = log/activity.log
+LOG_FILE_ACTIVE = out_application/log/activity.log
+LOG_FILE_ERROR  = out_application/log/error.log
 
 # ACTION
-FOLDER_STORAGE  = storage
-FOLDER_TMP      = tmp_file
-JSON_FILE_CHUNK = aux/chunk.json
-JSON_FILE_HASH  = aux/hash.json
+FOLDER_STORAGE  = out_application/storage
+FOLDER_TMP      = out_application/tmp_file
+JSON_FILE_CHUNK = out_application/collection/chunk.json
+JSON_FILE_HASH  = out_application/collection/hash.json
