+++ out_application/doc/state-action-clear.png
