+++ out_application/doc/state-action-clear.wsd
