+++ out_application/doc/state-action-remove.wsd
