+++ out_application/input_file_examples/day_book_17.07.txt
