+++ out_application/input_file_examples/day_book_18.07.txt
