+++ out_application/input_file_examples/day_book_19.07.txt
