+++ out_application/input_file_examples/day_book_20.07.txt
