+++ src/interfaces/action.go
@@ -1,0 +1,9 @@
+package interfaces
+
+type ActionBase interface {
+	FeatureUpload() error
+	FeatureDownload() error
+	FeatureRemove() error
+	FeatureClear() error
+	GetActionType() string
+}
